+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -2,8 +2,11 @@ package ru.javawebinar.topjava.repository.mock;
 
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.util.UserUtil;
 
+import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
@@ -17,8 +20,8 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        save(MealsUtil.MEALS.get(0), 0);
-        save(MealsUtil.MEALS.get(1), 0);
+        save(MealsUtil.MEALS.get(0), UserUtil.getUser().getId());
+        save(MealsUtil.MEALS.get(1), UserUtil.getAdmin().getId());
     }
 
     @Override
@@ -59,6 +62,17 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
         Map<Integer, Meal> meals = repository.get(userId);
         if (!meals.isEmpty()) {
             return meals.values().stream().
+                    sorted(Comparator.comparing(Meal::getDateTime).reversed()).collect(Collectors.toList());
+        }
+        else return Collections.EMPTY_LIST;
+    }
+
+    @Override
+    public Collection<Meal> getBetween(LocalDateTime startDT, LocalDateTime endDT, int userId) {
+        Map<Integer, Meal> meals = repository.get(userId);
+        if (!meals.isEmpty()) {
+            return meals.values().stream().
+                    filter(f -> DateTimeUtil.isBetween(f.getDateTime(), startDT, endDT)).
                     sorted(Comparator.comparing(Meal::getDateTime).reversed()).collect(Collectors.toList());
         }
         else return Collections.EMPTY_LIST;