+++ src/main/java/ru/javawebinar/topjava/service/MealService.java
@@ -1,4 +1,20 @@
 package ru.javawebinar.topjava.service;
 
+import ru.javawebinar.topjava.model.Meal;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
+
+import java.time.LocalDateTime;
+import java.util.Collection;
+
 public interface MealService {
+    Meal save(Meal meal, int userId) throws NotFoundException;
+
+    boolean delete(int id, int userId) throws NotFoundException;
+
+    Meal get(int id, int userId) throws NotFoundException;
+
+    Collection<Meal> getAll(int userId);
+
+    Collection<Meal> getBetween(LocalDateTime startDT, LocalDateTime endDT, int userId);
+
 }