+++ src/main/java/ru/javawebinar/topjava/util/UserUtil.java
@@ -12,4 +12,12 @@ public class UserUtil {
             new User(1, "Userok", "user@neskaz.hu", "password", Role.ROLE_USER),
             new User(2, "Odmin", "likeaboss@neskaz.hu", "P@s$w0r@", Role.ROLE_ADMIN)
     );
+
+    public static User getUser() {
+        return USERS.get(1);
+    }
+
+    public static User getAdmin() {
+        return USERS.get(2);
+    }
 }