+++ src/main/java/ru/javawebinar/topjava/web/UserServlet.java
@@ -15,7 +15,7 @@ public class UserServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        log.debug("forward to users");
+        log.debug("forward ru.javawebinar.topjava.to users");
         request.getRequestDispatcher("/users.jsp").forward(request, response);
     }
 }