+++ src/main/java/ru/javawebinar/topjava/web/meal/MealRestController.java
@@ -1,8 +1,15 @@
 package ru.javawebinar.topjava.web.meal;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
 import ru.javawebinar.topjava.service.MealService;
 
+@Controller
 public class MealRestController {
     private MealService service;
+
+    @Autowired
+    private MealService mealService;
+
 
 }